// customer_service/cs-frontend/src/app/models/detail.model.ts
export default class Detail {
  _buyId: string;
  _pdtId: string;
  _rptId: string;
  custUsername: string;
  saleUsername: string;
  emplUsername: string;
  itemname: string;
  brand: string;
  unitprice: number;
  quantity: number;
  orderDate: Date;
  deliverDate: Date;
  cancel: boolean;
  subject: string;
  complaint: string;
  status: boolean;
  resolution: string;
}